requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
prometheus-client==0.9.0
py-solaredge==0.0.4.8
requests==2.25.0
urllib3==1.26.2
Werkzeug==1.0.1
